An invalid query raises an Internal Server Error
An invalid query raises an Internal Server Error, but at the client
side I can't figure out what happened in the server.
The server should return an error code or the response should have an
return status indicating success or failure, and then we could have
many codes for failures.